IT Job Market ETL Pipeline
This repository contains the code for an ETL (Extract, Transform, Load) pipeline focused on collecting and analysing data related to the IT job market. The pipeline extracts information from a source, transforms it into a structured format, and loads it into a database for analysis. The project includes data cleaning, encoding detection, transformation, database integration, and data visualisation.